Bugfixes
In certain cases, e. g. with KPEnhancedEntryView installed, KeePass did not properly drop to background when you want to read an OTP QR code from screen. This is fixed now.

OTP QR codes are searched for max. 30 seconds and it was not possible to close the OTP setup window while the search is running. This is fixed now.

Closes #138